飞机大战3.6复仇风暴

+++ src/com/neuedu/base/BaseSprite.java
@@ -3,30 +3,30 @@ package com.neuedu.base;
 import java.awt.Rectangle;
 
 public abstract class BaseSprite {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public BaseSprite() {
     }
 
-    public BaseSprite(int x, int y) {
+    public BaseSprite(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
