+++ src/com/neuedu/runtime/Bullet.java
@@ -31,7 +31,7 @@ public class Bullet extends BaseSprite implements Moveable, Drawable {
     @Override
     public void draw(Graphics g) {
         move();
-        g.drawImage(image, getX(), getY(), image.getWidth(null), image.getHeight(null), null);
+        g.drawImage(image, (int) getX(), (int) getY(), image.getWidth(null), image.getHeight(null), null);
         outOfBound();
     }
 
@@ -50,7 +50,7 @@ public class Bullet extends BaseSprite implements Moveable, Drawable {
 
     @Override
     public Rectangle getRectangle() {
-        return new Rectangle(getX(), getY(), image.getWidth(null), image.getHeight(null));
+        return new Rectangle((int) getX(), (int) getY(), image.getWidth(null), image.getHeight(null));
     }
 
     public void collisionTest(List<EnemyPlane> enemyPlaneList, List<Boss> bossList, List<Stones> stonesList) {
