+++ src/com/neuedu/runtime/Stones.java
@@ -18,8 +18,8 @@ public class Stones extends BaseSprite implements Moveable, Drawable {
 
     private int speed = FrameConstant.GAME_SPEED;
 
-    public Stones(){
-        this(0,0, ImageMap.get("stone"),0);
+    public Stones() {
+        this(0, 0, ImageMap.get("stone"), 0);
     }
 
 
@@ -32,7 +32,7 @@ public class Stones extends BaseSprite implements Moveable, Drawable {
     @Override
     public void draw(Graphics g) {
         move();
-        g.drawImage(image, getX(), getY(), image.getWidth(null), image.getHeight(null), null);
+        g.drawImage(image, (int) getX(), (int) getY(), image.getWidth(null), image.getHeight(null), null);
         outOfBound();
 
     }
@@ -50,17 +50,17 @@ public class Stones extends BaseSprite implements Moveable, Drawable {
         }
     }
 
-    public Rectangle getRectangle(){
-        return new Rectangle(getX(), getY(), image.getWidth(null) / 10 * 9, image.getHeight(null) / 10 * 9);
+    public Rectangle getRectangle() {
+        return new Rectangle((int) getX(), (int) getY(), image.getWidth(null) / 10 * 9, image.getHeight(null) / 10 * 9);
     }
 
-    public void collisionTest(Plane plane,Plane2 plane2){
+    public void collisionTest(Plane plane, Plane2 plane2) {
 //        撞到陨石
         GameFrame gameFrame = DataStore.get("gameFrame");
-        if(plane.getRectangle().intersects(this.getRectangle())){
+        if (plane.getRectangle().intersects(this.getRectangle())) {
             plane.hp--;
         }
-        if(plane2.getRectangle().intersects(this.getRectangle())){
+        if (plane2.getRectangle().intersects(this.getRectangle())) {
             plane2.hp--;
         }
